Rename wwnet to WWNet
