+++ Code/WWNet/netutil.cpp
