+++ Code/WWNet/networkobject.cpp
