+++ Code/WWNet/networkobject.h
